(BSR) fix(gsheet): fix fraudulent device sheet ID typo & refactor microservice structure (#5355)
* feat(gsheet): add logging

* chore(dependency): bump packages

* fix(gsheet): typo fraudulent device sheet ID

* refactor(gsheet): separation of concerns
